requirements.txt
Django
djangorestframework
python-decouple
psycopg[binary]
dj-rest-auth[with_social]
django-cors-headers